requirements.txt
asgiref==3.4.1
Django==4.0b1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.4
utils==1.0.1
